+++ extensions/Worker.Extensions.Storage.Blobs/src/Constants.cs
@@ -7,9 +7,6 @@ namespace Microsoft.Azure.Functions.Worker.Extensions.Storage.Blobs
     {
         internal const string Storage = "Storage";
         internal const string BlobExtensionName = "AzureStorageBlobs";
-        internal const string Connection = "Connection";
-        internal const string ContainerName = "ContainerName";
-        internal const string BlobName = "BlobName";
 
         // Media content types
         internal const string JsonContentType = "application/json";