+++ samples/WorkerBindingSamples/Blob/BlobTriggerBindingSamples.cs
@@ -8,6 +8,9 @@ using Microsoft.Extensions.Logging;
 
 namespace SampleApp
 {
+    /// <summary>
+    /// Samples demonstrating binding to the types supported by the `BlobTrigger` binding.
+    /// </summary>
     public class BlobTriggerBindingSamples
     {
         private readonly ILogger<BlobTriggerBindingSamples> _logger;
@@ -17,6 +20,14 @@ namespace SampleApp
             _logger = logger;
         }
 
+        /// <summary>
+        /// This sample demonstrates how to retrieve the contents of a blob file when a blob
+        /// is added or updated in the given container. The code uses a <see cref="BlobClient"/>
+        /// instance to read contents of the blob. The string {name} in the blob trigger path
+        /// creates a binding expression that you can use in function code to access the file
+        /// name of the triggering blob.
+        /// The <see cref="BlobClient"/> instance could also be used for write operations.
+        /// </summary>
         [Function(nameof(BlobClientFunction))]
         public async Task BlobClientFunction(
             [BlobTrigger("client-trigger/{name}")] BlobClient client, string name)
@@ -26,6 +37,12 @@ namespace SampleApp
             _logger.LogInformation("Blob name: {name} -- Blob content: {content}", name, content);
         }
 
+        /// <summary>
+        /// This sample demonstrates how to retrieve the contents of a blob file when a blob
+        /// is added or updated in the given container by binding to a <see cref="Stream"/>.
+        /// The string {name} in the blob trigger path creates a binding expression that you
+        /// can use in function code to access the file name of the triggering blob.
+        /// </summary>
         [Function(nameof(BlobStreamFunction))]
         public async Task BlobStreamFunction(
             [BlobTrigger("stream-trigger/{name}")] Stream stream, string name)
@@ -35,6 +52,10 @@ namespace SampleApp
             _logger.LogInformation("Blob name: {name} -- Blob content: {content}", name, content);
         }
 
+        /// <summary>
+        /// This sample demonstrates how to retrieve the contents of a blob file when a blob
+        /// is added or updated in the given container by binding to a <see cref="Byte[]"/>.
+        /// </summary>
         [Function(nameof(BlobByteArrayFunction))]
         public void BlobByteArrayFunction(
             [BlobTrigger("byte-trigger")] Byte[] data)
@@ -42,6 +63,10 @@ namespace SampleApp
             _logger.LogInformation("Blob content: {content}", Encoding.Default.GetString(data));
         }
 
+        /// <summary>
+        /// This sample demonstrates how to retrieve the contents of a blob file when a blob
+        /// is added or updated in the given container by binding to a <see cref="string"/>.
+        /// </summary>
         [Function(nameof(BlobStringFunction))]
         public void BlobStringFunction(
             [BlobTrigger("string-trigger")] string data)
@@ -49,6 +74,11 @@ namespace SampleApp
             _logger.LogInformation("Blob content: {content}", data);
         }
 
+        /// <summary>
+        /// This sample demonstrates how to retrieve the contents of a blob file when a blob
+        /// is added or updated in the given container by binding to a <see cref="Book"/> (POCO).
+        /// The content of the blob must be JSON deserializable into the type of the parameter.
+        /// </summary>
         [Function(nameof(BlobBookFunction))]
         public void BlobBookFunction(
             [BlobTrigger("book-trigger")] Book data)