+++ test/FunctionMetadataGeneratorTests/FunctionMetadataGeneratorTests.cs
@@ -171,7 +171,7 @@ namespace Microsoft.Azure.Functions.SdkTests
 
             AssertDictionary(extensions, new Dictionary<string, string>
             {
-                { "Microsoft.Azure.WebJobs.Extensions.Storage.Queues", "5.1.3" },
+                { "Microsoft.Azure.WebJobs.Extensions.Storage.Queues", "5.1.2" },
                 { "Microsoft.Azure.WebJobs.Extensions.Storage.Blobs", "5.1.3" },
             });
 
@@ -208,10 +208,8 @@ namespace Microsoft.Azure.Functions.SdkTests
                 {
                     { "Name", "blobinput" },
                     { "Type", "blob" },
-                    { "DataType", "String"},
                     { "Direction", "In" },
                     { "blobPath", "container2" },
-                    { "Cardinality", "Many" },
                     { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding" , "True"} } }
                 });
             }
@@ -227,7 +225,6 @@ namespace Microsoft.Azure.Functions.SdkTests
                     { "Name", "blob" },
                     { "Type", "blobTrigger" },
                     { "Direction", "In" },
-                    { "DataType", "String"},
                     { "path", "container2/%file%" },
                     { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding" , "True"} } }
                 });
@@ -321,7 +318,6 @@ namespace Microsoft.Azure.Functions.SdkTests
                     { "Type", "blob" },
                     { "Direction", "In" },
                     { "blobPath", "container2/%file%" },
-                    { "Cardinality", "One" },
                     { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding" , "True"} } }
                 });
             }
@@ -371,7 +367,6 @@ namespace Microsoft.Azure.Functions.SdkTests
                     { "Type", "blob" },
                     { "Direction", "In" },
                     { "blobPath", "container2" },
-                    { "Cardinality", "Many" },
                     { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding", "True" } } }
                 });
             }
@@ -391,8 +386,6 @@ namespace Microsoft.Azure.Functions.SdkTests
                     { "Type", "blob" },
                     { "Direction", "In" },
                     { "blobPath", "container2" },
-                    { "Cardinality", "Many" },
-                    { "DataType", "String" },
                     { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding", "True" } } }
                 });
             }
@@ -487,7 +480,7 @@ namespace Microsoft.Azure.Functions.SdkTests
 
             AssertDictionary(extensions, new Dictionary<string, string>
             {
-                { "Microsoft.Azure.WebJobs.Extensions.Storage.Queues", "5.1.3" },
+                { "Microsoft.Azure.WebJobs.Extensions.Storage.Queues", "5.1.2" },
                 { "Microsoft.Azure.WebJobs.Extensions.Storage.Blobs", "5.1.3" },
             });
 
@@ -939,7 +932,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
             public object BlobToBlobs(
                 [BlobTrigger("container2/%file%")] string blob,
-                [BlobInput("container2", IsBatched = true)] IEnumerable<string> blobinput)
+                [BlobInput("container2")] IEnumerable<string> blobinput)
             {
                 throw new NotImplementedException();
             }
@@ -955,7 +948,6 @@ namespace Microsoft.Azure.Functions.SdkTests
             {
                 throw new NotImplementedException();
             }
-
 
             [Function("BlobClientToBlobStringFunction")]
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
@@ -1000,7 +992,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
             public object BlobStringToBlobStringArray(
                 [BlobTrigger("container2/%file%")] string blob,
-                [BlobInput("container2", IsBatched = true)] string[] blobinput)
+                [BlobInput("container2")] string[] blobinput)
             {
                 throw new NotImplementedException();
             }
@@ -1010,7 +1002,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
             public object BlobStringToBlobClientEnumerable(
                 [BlobTrigger("container2/%file%")] string blob,
-                [BlobInput("container2", IsBatched = true)] IEnumerable<BlobClient> blobinput)
+                [BlobInput("container2")] IEnumerable<BlobClient> blobinput)
             {
                 throw new NotImplementedException();
             }
@@ -1019,7 +1011,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
             public object BlobStringToBlobPocoEnumerable(
                 [BlobTrigger("container2/%file%")] string blob,
-                [BlobInput("container2", IsBatched = true)] IEnumerable<Poco> blobinput)
+                [BlobInput("container2")] IEnumerable<Poco> blobinput)
             {
                 throw new NotImplementedException();
             }
@@ -1028,7 +1020,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
             public object BlobStringToBlobPocoArray(
                 [BlobTrigger("container2/%file%")] string blob,
-                [BlobInput("container2", IsBatched = true)] Poco[] blobinput)
+                [BlobInput("container2")] Poco[] blobinput)
             {
                 throw new NotImplementedException();
             }