+++ extensions/Worker.Extensions.Storage.Blobs/release_notes.md
@@ -4,7 +4,8 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.Storage.Blobs <version>
+### Microsoft.Azure.Functions.Worker.Extensions.Storage.Blobs 6.0.0
 
-- Add support for SDK-type bindings via deferred binding feature #1108
-- Assign cardinality correctly for Blob collection scenarios #1271
+- Add support for SDK-type bindings via deferred binding feature
+- Remove IsBatched property from BlobInput binding attribute
+- Infer binding cardinality based on blobPath